Be more lenient on the dependencies
This project isn't doing anything particularly fancy meaning that we don't need
to depend on the bleeding edge versions of our requirements i.e. we can use the
slightly older versions of scikit-learn, pandas and scipy that are available
pre-compiled in the package repositories of many Linux distributions. This makes
the project easier to set up and therefore more approachable.